pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "terrabot"
version = "0.1.0"
description = "AI-powered Terraform automation platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "typer[all]>=0.15.0",
    "rich>=13.9.0",
    "python-hcl2>=6.1.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "anthropic>=0.45.0",
    "jinja2>=3.1.0",
    "httpx>=0.28.0",
    "sse-starlette>=2.2.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
terrabot = "backend.cli.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test-*.py", "test_*.py"]
python_classes = ["Test"]
python_functions = ["test_"]
markers = [
    "unit: fast unit tests with no external dependencies",
    "integration: tests that exercise the full API stack",
    "e2e: end-to-end tests requiring a running server and terraform binary",
]
